Added HomePage load test
