+++ resources/testng-WA.xml
@@ -20,6 +20,7 @@
 			</run>
 		</groups>
 		<classes>
+			<class name="gov.nci.webanalyticstests.load.HomePage_Test" />
 			<class name="gov.nci.webanalyticstests.load.PageLoad_Test" />
 			<class name="gov.nci.webanalyticstests.click.WaBasicSearch_Test" />
 			<class name="gov.nci.webanalyticstests.click.SitewideSearch_Test" />
